Add variablem

+++ OneDrive/Desktop/148229 -Christopher Muturi/calculator.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    double num1, num2;
+    float num1, num2;
     
     cout << "=== BASIC CALCULATOR ===" << endl;
     cout << "Enter first number: ";
